#298: Implement temporary ItemStoreMother
