+++ Frontend/ShoppingList.Frontend.Redux.Tests/Items/States/AvailableTests.cs
@@ -1,6 +1,6 @@
 ﻿using FluentAssertions;
 using ProjectHermes.ShoppingList.Frontend.Redux.Items.States;
-using ProjectHermes.ShoppingList.Frontend.Redux.TestKit.Common;
+using ProjectHermes.ShoppingList.Frontend.Redux.TestKit.Items.States;
 
 namespace ProjectHermes.ShoppingList.Frontend.Redux.Tests.Items.States;
 
@@ -84,14 +84,8 @@ public class AvailableTests
         {
             Stores = new List<ItemStore>
             {
-                new DomainTestBuilder<ItemStore>().Create() with
-                {
-                    Sections = CreateSections().ToList()
-                },
-                new DomainTestBuilder<ItemStore>().Create() with
-                {
-                    Sections = CreateSections().ToList()
-                }
+                ItemStoreMother.GetValid(),
+                ItemStoreMother.GetValid()
             };
         }
 
@@ -119,19 +113,6 @@ public class AvailableTests
                     .Select(s => new ItemStoreSection(s.Id, s.Name, s.IsDefaultSection, s.SortingIndex))
                     .ToList()
             };
-        }
-
-        private static IEnumerable<ItemStoreSection> CreateSections()
-        {
-            for (int i = 0; i < 2; i++)
-            {
-                var section = new DomainTestBuilder<ItemStoreSection>().Create() with
-                {
-                    IsDefaultSection = i == 0,
-                    SortingIndex = i
-                };
-                yield return section;
-            }
         }
     }
